+++ CHANGELOG.md
@@ -1,6 +1,43 @@
 # Changelog
 
 
+## v0.0.4
+
+[compare changes](https://github.com/haus23/tipprunde/compare/v0.0.3...v0.0.4)
+
+### 🚀 Enhancements
+
+- Add logo component. ([3fe6e98](https://github.com/haus23/tipprunde/commit/3fe6e98))
+- Add dark/light-mode idea. ([64dd198](https://github.com/haus23/tipprunde/commit/64dd198))
+- Create custom icon component. ([3f46be4](https://github.com/haus23/tipprunde/commit/3f46be4))
+- Add button component. ([38012b1](https://github.com/haus23/tipprunde/commit/38012b1))
+- Add popover component. ([298a4e3](https://github.com/haus23/tipprunde/commit/298a4e3))
+- Create a preference session and load it in the root loader. ([6f0ca67](https://github.com/haus23/tipprunde/commit/6f0ca67))
+- Add action to set colorScheme preference in the pref cookie. ([4463499](https://github.com/haus23/tipprunde/commit/4463499))
+- Add theme icons. ([b33ffce](https://github.com/haus23/tipprunde/commit/b33ffce))
+- Add menu component. ([d552943](https://github.com/haus23/tipprunde/commit/d552943))
+- Add client component and hook to switch colorScheme. ([25907b7](https://github.com/haus23/tipprunde/commit/25907b7))
+
+### 🩹 Fixes
+
+- Remove application error. Temporarily no route meta data. ([#10](https://github.com/haus23/tipprunde/pull/10))
+- Workaround to set default border color. ([4cd1676](https://github.com/haus23/tipprunde/commit/4cd1676))
+
+### 💅 Refactors
+
+- Align future tv styles naming. ([1088d50](https://github.com/haus23/tipprunde/commit/1088d50))
+- Change dark-mode selector to class-name. ([e7377f5](https://github.com/haus23/tipprunde/commit/e7377f5))
+- Use loader-data and set class-name. ([d702763](https://github.com/haus23/tipprunde/commit/d702763))
+
+### 🏡 Chore
+
+- Use my preferred semantic versioning. ([6b34922](https://github.com/haus23/tipprunde/commit/6b34922))
+- Add react-aria-components library. ([f66fa6a](https://github.com/haus23/tipprunde/commit/f66fa6a))
+
+### ❤️ Contributors
+
+- Micha Buchholz <micha@haus23.net>
+
 ## v0.0.3
 
 [compare changes](https://github.com/haus23/tipprunde/compare/v0.0.2...v0.0.3)